Refactor error handling and logging in hazard prediction module
- Replaced `sprintf` with `snprintf` in multiple files to enhance safety and prevent buffer overflows in error messages.
- Updated logging messages in `hazverr`, `obsverr`, `prederr`, and `obsloop` to improve clarity and consistency.
- Modified `initprz` to use `snprintf` for constructing version strings.
- Changed variable handling in `opnfils` to improve clarity and maintainability.
- Updated Makefile configurations across various directories to use `--foreign` instead of `--gnu` for Automake.
- Enhanced `ylwrap` script for better error handling and output management.
- Added a new test script `run_local_example.sh` to facilitate local testing of the hazard and hazpred executables with fixture files.

+++ src/hazpred/alocmem.c
@@ -8,7 +8,7 @@
 #include "hzfxit.h"
 
 void alocmem(void){
-  int i;
+  size_t i;
 
   data_xv = hzf_memget(infilect*sizeof(struct xvgetstr));
   inhaz_xv = hzf_memget(nvars*sizeof(struct xvgetstr));
@@ -48,6 +48,6 @@ void alocmem(void){
   memcpy(names[6],"ALPHA   ",8);
   memcpy(names[7],"ETA     ",8);
   memcpy(names[8],"E0      ",8);
-  for(i=0; i<(C->Ntheta*C->Ntheta); i++)
+  for(i=0; i<(size_t)(C->Ntheta*C->Ntheta); i++)
     C->cov[i] = ZERO;
 }
